+++ src/cpp/flann/nn/index_testing.h
@@ -45,7 +45,7 @@
 namespace flann
 {
 
-inline int countCorrectMatches(int* neighbors, int* groundTruth, int n)
+inline int countCorrectMatches(size_t* neighbors, size_t* groundTruth, int n)
 {
     int count = 0;
     for (int i=0; i<n; ++i) {
@@ -62,7 +62,7 @@ inline int countCorrectMatches(int* neighbors, int* groundTruth, int n)
 
 template <typename Distance>
 typename Distance::ResultType computeDistanceRaport(const Matrix<typename Distance::ElementType>& inputData, typename Distance::ElementType* target,
-                                                    int* neighbors, int* groundTruth, int veclen, int n, const Distance& distance)
+		size_t* neighbors, size_t* groundTruth, int veclen, int n, const Distance& distance)
 {
     typedef typename Distance::ResultType DistanceType;
 
@@ -84,7 +84,7 @@ typename Distance::ResultType computeDistanceRaport(const Matrix<typename Distan
 
 template <typename Index, typename Distance>
 float search_with_ground_truth(Index& index, const Matrix<typename Distance::ElementType>& inputData,
-                               const Matrix<typename Distance::ElementType>& testData, const Matrix<int>& matches, int nn, int checks,
+                               const Matrix<typename Distance::ElementType>& testData, const Matrix<size_t>& matches, int nn, int checks,
                                float& time, typename Distance::ResultType& dist, const Distance& distance, int skipMatches)
 {
     typedef typename Distance::ElementType ElementType;
@@ -97,13 +97,13 @@ float search_with_ground_truth(Index& index, const Matrix<typename Distance::Ele
 
     SearchParams searchParams(checks);
 
-    int* indices = new int[nn+skipMatches];
+    size_t* indices = new size_t[nn+skipMatches];
     DistanceType* dists = new DistanceType[nn+skipMatches];
     
-    Matrix<int> indices_mat(indices, 1, nn+skipMatches);
+    Matrix<size_t> indices_mat(indices, 1, nn+skipMatches);
     Matrix<DistanceType> dists_mat(dists, 1, nn+skipMatches);
         
-    int* neighbors = indices + skipMatches;
+    size_t* neighbors = indices + skipMatches;
 
     int correct = 0;
     DistanceType distR = 0;
@@ -140,7 +140,7 @@ float search_with_ground_truth(Index& index, const Matrix<typename Distance::Ele
 
 template <typename Index, typename Distance>
 float test_index_checks(Index& index, const Matrix<typename Distance::ElementType>& inputData,
-                        const Matrix<typename Distance::ElementType>& testData, const Matrix<int>& matches,
+                        const Matrix<typename Distance::ElementType>& testData, const Matrix<size_t>& matches,
                         int checks, float& precision, const Distance& distance, int nn = 1, int skipMatches = 0)
 {
     typedef typename Distance::ResultType DistanceType;
@@ -157,7 +157,7 @@ float test_index_checks(Index& index, const Matrix<typename Distance::ElementTyp
 
 template <typename Index, typename Distance>
 float test_index_precision(Index& index, const Matrix<typename Distance::ElementType>& inputData,
-                           const Matrix<typename Distance::ElementType>& testData, const Matrix<int>& matches,
+                           const Matrix<typename Distance::ElementType>& testData, const Matrix<size_t>& matches,
                            float precision, int& checks, const Distance& distance, int nn = 1, int skipMatches = 0)
 {
     typedef typename Distance::ResultType DistanceType;
